fix: update package exports
