+++ .github/copilot-instructions.md
@@ -1,9 +1,3 @@
-dont add any comments to the code, no docstrings, no explanations, just the code
-use best practices, use type hints, use imports from the same package, do not import unused modules, do not import unused functions
-check for repetitions, do not repeat code, do not repeat functions, reuse code and functions
-use best patterns if applicable
-
-to run python use uv command like
-uv run -m main.py
-uv add something
-etc
+This repository uses a single package `glopt` under `src/`.
+Place all source files inside this package and prefer relative imports within it.
+Outputs from experiments are stored in the `runs/` directory.
