+++ README.md
@@ -4,8 +4,7 @@ Simple application for exploring cost-efficient license allocation in social net
 
 ## Repository structure
 
-- `src/` – lightweight Python package with models, algorithms and helpers
-- `scripts/` – runnable scripts used for experiments and demos
+- `src/glopt/` – project package with models, algorithms, I/O and CLI helpers
 - `Makefile` – helper targets for common tasks (`test`, `benchmark`, `dynamic`, `format`, `lint`, `install`)
 
 ## Quick start
@@ -16,4 +15,4 @@ make test         # execute unit tests verifying algorithm quality
 make dynamic      # execute dynamic network simulation
 ```
 
-Simulation and benchmark outputs are stored in the `results/` directory.
+Simulation and benchmark outputs are stored in the `runs/` directory.
