+++ src/glopt/algorithms/naive.py
@@ -3,8 +3,8 @@ from typing import Any, Iterator, List, Sequence, Set, Tuple
 
 import networkx as nx
 
-from src import Algorithm, LicenseGroup, LicenseType, Solution
-from src.solution_builder import SolutionBuilder
+from ..core import Algorithm, LicenseGroup, LicenseType, Solution
+from ..core.solution_builder import SolutionBuilder
 
 
 Assignment = List[Tuple[LicenseType, Any, Set[Any]]]  # [(license, owner, members_without_owner), ...]
