+++ src/glopt/algorithms/tabu_search.py
@@ -2,9 +2,9 @@ from collections import deque
 from typing import Any, Deque, List
 import networkx as nx
 
-from src import Algorithm, LicenseType, Solution
-from src.solution_validator import SolutionValidator
-from src.mutations import MutationOperators
+from ..core import Algorithm, LicenseType, Solution
+from ..core.solution_validator import SolutionValidator
+from ..core.mutations import MutationOperators
 from .greedy import GreedyAlgorithm
