+++ src/glopt/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Command-line entry points for glopt."""
