+++ src/glopt/io/csv_writer.py
@@ -1,11 +1,27 @@
 import csv
 import os
+from dataclasses import asdict
 from datetime import datetime
-from typing import Dict, Any
+from typing import Any, Dict, Iterable
+
+
+def write_csv(csv_dir: str, run_id: str, rows: Iterable[Any]) -> str:
+    out_path = os.path.join(csv_dir, f"{run_id}.csv")
+    first = True
+    with open(out_path, "w", newline="", encoding="utf-8") as f:
+        writer = None
+        for r in rows:
+            d = asdict(r)
+            if first:
+                writer = csv.DictWriter(f, fieldnames=list(d.keys()))
+                writer.writeheader()
+                first = False
+            writer.writerow(d)  # type: ignore[union-attr]
+    return out_path
 
 
 class BenchmarkCSVWriter:
-    def __init__(self, output_dir: str = "results/stats"):
+    def __init__(self, output_dir: str = "runs/stats"):
         self.output_dir = output_dir
         os.makedirs(output_dir, exist_ok=True)
         timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
