+++ src/glopt/io/graph_generator.py
